1.部门删除控制逻辑添加 2.file页面复制链接修改

+++ bootdo/src/main/java/com/bootdo/system/controller/DeptController.java
@@ -108,8 +108,17 @@ public class DeptController extends BaseController {
 		if (Constant.DEMO_ACCOUNT.equals(getUsername())) {
 			return R.error(1, "演示系统不允许修改,完整体验请部署程序");
 		}
-		if (sysDeptService.remove(deptId) > 0) {
-			return R.ok();
+		Map<String, Object> map = new HashMap<String, Object>();
+		map.put("parentId", deptId);
+		if(sysDeptService.count(map)>0) {
+			return R.error(1, "包含下级部门,不允许修改");
+		}
+		if(sysDeptService.checkDeptHasUser(deptId)) {
+			if (sysDeptService.remove(deptId) > 0) {
+				return R.ok();
+			}
+		}else {
+			return R.error(1, "部门包含用户,不允许修改");
 		}
 		return R.error();
 	}
